--- a/lib/plugins/helper/toc.js
+++ b/lib/plugins/helper/toc.js
@@ -1,15 +1,21 @@
 'use strict';
 
-let cheerio;
 const { escapeHTML } = require('hexo-util');
+const { DomHandler, DomUtils, Parser } = require('htmlparser2');
+
+const parseHtml = (html) => {
+  const handler = new DomHandler(null, {});
+  new Parser(handler, {}).end(html);
+  return handler.dom;
+};
 
 function tocHelper(str, options = {}) {
-  if (!cheerio) cheerio = require('cheerio');
+  const dom = parseHtml(str);
 
-  const $ = cheerio.load(str);
   const headingsMaxDepth = Object.prototype.hasOwnProperty.call(options, 'max_depth') ? options.max_depth : 6;
   const headingsSelector = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'].slice(0, headingsMaxDepth).join(',');
-  const headings = $(headingsSelector);
+
+  const headings = DomUtils.find(el => headingsSelector.includes(el.tagName), dom, true);
 
   if (!headings.length) return '';
 
@@ -21,15 +27,15 @@ function tocHelper(str, options = {}) {
   let lastLevel = 0;
 
   function getId(ele) {
-    const id = $(ele).attr('id');
-    const $parent = $(ele).parent();
-    return id || ($parent.length < 1 ? null : getId($parent));
+    const { id } = ele.attribs;
+    const { parent } = ele;
+    return id || (parent.length < 1 ? null : getId(parent));
   }
 
-  headings.each(function() {
-    const level = +this.name[1];
-    const id = getId(this);
-    const text = escapeHTML($(this).text());
+  for (const el of headings) {
+    const level = +el.name[1];
+    const id = getId(el);
+    const text = escapeHTML(DomUtils.getText(el));
 
     lastNumber[level - 1]++;
 
@@ -67,7 +73,7 @@ function tocHelper(str, options = {}) {
     result += `<span class="${className}-text">${text}</span></a>`;
 
     lastLevel = level;
-  });
+  }
 
   for (let i = firstLevel - 1; i < lastLevel; i++) {
     result += '</li></ol>';