<!--
Thank you for creating a pull request to contribute to Hexo code! Before you open the request please answer the following questions to help it be more easily integrated. Please check the boxes "[ ]" with "[x]" when done too.
-->

## What does it do?

This PR is the last pieces of #3677. We finally made it.

## How to test

```sh
git clone -b htmlparser2-toc https://github.com/sukkaw/hexo.git
cd hexo
npm install
npm test
```

## Screenshots

Since `hexo-theme-landscape` is not using toc helper, I have set up a benchmark using my `hexo-theme-suka`, with hexo's `meta_generator` and `external_link` disabled.

**Node.js 8**

| | Process Time | Memory  |
|:---:|:---:|:---:|
| Current Master Branch | 21s | 528MB |
| This PR | 17s | 421MB |

**Nodejs.10**

| | Process Time | Memory  |
|:---:|:---:|:---:|
| Current Master Branch | 20s | 527MB |
| This PR | 17s | 464MB |

**Nodejs.12**

| | Process Time | Memory  |
|:---:|:---:|:---:|
| Current Master Branch | 24s | 460MB |
| This PR | 22s | 470MB |

## Pull request tasks

- [ ] Add test cases for the changes.
- [x] Passed the CI test.