@SukkaW 2019-11-18
This PR should be superseded by utilizing newer version of `hexo-util` after https://github.com/hexojs/hexo-util/pull/137 is merged and released in `hexo-util@1.6.0`.