@curbengh 2019-12-07
> This PR should be superseded by utilizing newer version of `hexo-util` after [hexojs/hexo-util#137](https://github.com/hexojs/hexo-util/pull/137).

Can this PR use https://github.com/hexojs/hexo-util/pull/137 branch temporarily for testing? Once that PR is released, just need to update hexo-util in this PR.