@curbengh 2019-12-19
[So far](https://github.com/hexojs/hexo/compare/4.1.1...master) we haven't merge any new feature yet, perhaps this can be released into 4.1.2?