@SukkaW 2019-12-20
@curbengh `min_depth` could be bring up along with `tocObj()` so there won't be only one feature.