@curbengh 2019-12-20
> `min_depth` could be bring up along with `tocObj()` so there won't be only one feature.

That makes two features then. Sufficient for 4.2.0. #3891 can be deferred to 4.3+.